These are the easy parts with no conflicts.  This is most of the libc stuff, but not all of it.  There will be more to follow.

+++ otter/libc/string.h
@@ -4,35 +4,35 @@
 #include <stddef.h>
 
 /* Copying (7.21.2) */
-void* memcpy(void* s1, const void* s2, int n);
-void* memmove(void* s1, const void* s2, int n);
+void* memcpy(void* s1, const void* s2, size_t n);
+void* memmove(void* s1, const void* s2, size_t n);
 char* strcpy(char* s1, const char* s2);
-char* strncpy(char* s1, const char* s2, int n);
+char* strncpy(char* s1, const char* s2, size_t n);
 
 /* Concatination (7.21.3) */
 char* strcat(char* s1, const char* s2);
-char* strncat(char* s1, const char* s2, int n);
+char* strncat(char* s1, const char* s2, size_t n);
 
 /* Comparison (7.21.4) */
-int memcmp(const void* s1, const void* s2, int n);
+int memcmp(const void* s1, const void* s2, size_t n);
 int strcmp(const char* s1, const char* s2);
 int strcoll(const char* s1, const char* s2);
 int strxfrm(char* s1, const char* s2, int n);
 
 /* Search (7.21.5) */
-void* memchr(const void* s, int c, int n);
+void* memchr(const void* s, int c, size_t n);
 char* strchr(const char* s, int c);
-int strcspn(const char* s1, const char* s2);
+size_t strcspn(const char* s1, const char* s2);
 char* strpbrk(const char* s1, const char* s2);
 char* strrchr(const char* s, int c);
-int strspn(const char* s1, const char* s2);
+size_t strspn(const char* s1, const char* s2);
 char* strstr(const char* s1, const char* s2);
 char* strtok(char* s1, const char* s2);
 
 /* Misc (7.21.6) */
-void* memset(void* s, int c, int n);
+void* memset(void* s, int c, size_t n);
 char* strerror(int errnum);
-int strlen(const char* s);
+size_t strlen(const char* s);
 
 
 #endif
